+++ contracts/Liquidity.sol
@@ -1,4 +1,3 @@
-
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8;
 
@@ -6,61 +5,73 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "./Uniswap.sol";
 import "hardhat/console.sol";
 
-
+/// @title Liquidity Contract for adding and removing liqudity from Uniswap
+/// @author Ikhlas
+/// @notice The contract does not have the front end implemented
+/// @dev All function calls are currently implemented without side effects
+/// @custom:experimental This is an experimental contract.
 contract Liquidity {
-  address private constant FACTORY = 0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f;
-  address private constant ROUTER = 0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D;
-  address private constant WETH = 0xc778417E063141139Fce010982780140Aa0cD5Ab;
+    address private constant FACTORY =
+        0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f;
+    address private constant ROUTER =
+        0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D;
+    address private constant WETH = 0xc778417E063141139Fce010982780140Aa0cD5Ab;
 
-  event Log(string message, uint val);
+    /// @notice Events
+    event Log(string message, uint256 val);
 
-  function addLiquidity(
-    address _tokenA,
-    address _tokenB,
-    uint _amountA,
-    uint _amountB
-  ) external {
-    IERC20(_tokenA).transferFrom(msg.sender, address(this), _amountA);
-    IERC20(_tokenB).transferFrom(msg.sender, address(this), _amountB);
+    /// @notice Function to add liqudity to XXXETH Uniswap Pool
+    function addLiquidity(
+        address _tokenA,
+        address _tokenB,
+        uint256 _amountA,
+        uint256 _amountB
+    ) external {
+        IERC20(_tokenA).transferFrom(msg.sender, address(this), _amountA);
+        IERC20(_tokenB).transferFrom(msg.sender, address(this), _amountB);
 
-    IERC20(_tokenA).approve(ROUTER, _amountA);
-    IERC20(_tokenB).approve(ROUTER, _amountB);
+        IERC20(_tokenA).approve(ROUTER, _amountA);
+        IERC20(_tokenB).approve(ROUTER, _amountB);
 
-    (uint amountA, uint amountB, uint liquidity) =
-      IUniswapV2Router(ROUTER).addLiquidity(
-        _tokenA,
-        _tokenB,
-        _amountA,
-        _amountB,
-        1,
-        1,
-        msg.sender,
-        block.timestamp
-      );
+        (
+            uint256 amountA,
+            uint256 amountB,
+            uint256 liquidity
+        ) = IUniswapV2Router(ROUTER).addLiquidity(
+                _tokenA,
+                _tokenB,
+                _amountA,
+                _amountB,
+                1,
+                1,
+                msg.sender,
+                block.timestamp
+            );
 
-    emit Log("amountA", amountA);
-    emit Log("amountB", amountB);
-    emit Log("liquidity", liquidity);
-  }
+        emit Log("amountA", amountA);
+        emit Log("amountB", amountB);
+        emit Log("liquidity", liquidity);
+    }
 
-  function removeLiquidity(address _tokenA, address _tokenB) external {
-    address pair = IUniswapV2Factory(FACTORY).getPair(_tokenA, _tokenB);
+    /// @notice Function to remove liqudity from XXXETH Uniswap Pool
+    function removeLiquidity(address _tokenA, address _tokenB) external {
+        address pair = IUniswapV2Factory(FACTORY).getPair(_tokenA, _tokenB);
 
-    uint liquidity = IERC20(pair).balanceOf(msg.sender);
-    IERC20(pair).approve(ROUTER, liquidity);
+        uint256 liquidity = IERC20(pair).balanceOf(msg.sender);
+        IERC20(pair).approve(ROUTER, liquidity);
 
-    (uint amountA, uint amountB) =
-      IUniswapV2Router(ROUTER).removeLiquidity(
-        _tokenA,
-        _tokenB,
-        liquidity,
-        1,
-        1,
-        msg.sender,
-        block.timestamp
-      );
+        (uint256 amountA, uint256 amountB) = IUniswapV2Router(ROUTER)
+            .removeLiquidity(
+                _tokenA,
+                _tokenB,
+                liquidity,
+                1,
+                1,
+                msg.sender,
+                block.timestamp
+            );
 
-    emit Log("amountA", amountA);
-    emit Log("amountB", amountB);
-  }
+        emit Log("amountA", amountA);
+        emit Log("amountB", amountB);
+    }
 }
